backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
PyMuPDF==1.25.1
PyJWT==2.10.1
python-multipart==0.0.20
